compiler: clang: Fix -Wdefault-const-init-var-unsafe warnings
This commit addresses several instances of the '-Wdefault-const-init-var-unsafe'
and '-Wdefault-const-init-field-unsafe' warnings when building the kernel
with Clang 21. These warnings are triggered when `typeof` within kernel
macros (such as `time_before`, `time_after`, and `u64_to_user_ptr`)
implicitly declares a `const` variable or when a structure with a `const`
member is declared without an explicit initializer, which Clang considers unsafe.

The affected patterns are primarily:
1. Using `time_before()` or `time_after()` with a `const unsigned long`
   variable.
2. Using `u64_to_user_ptr()` with a `const unsigned long long` field
   from a structure.
3. Default initialization of a struct containing `const` members.

To resolve these warnings, the following changes are applied:
- Structures with `const` members are explicitly zero-initialized (e.g., `struct_name var = {};`).
- Arguments passed to `time_before()` and `time_after()` that are
  `const unsigned long` are explicitly cast to `unsigned long`.
- Arguments passed to `u64_to_user_ptr()` that are `const unsigned long long`
  are explicitly cast to `u64`.

These changes ensure that any implicit temporary variables created by `typeof`
within the macros are properly initialized, and that structures with `const`
members have a defined initial state. This satisfies Clang's stricter type
checking and eliminates the warnings without altering the intended kernel
behavior.

Affected files and lines include:
- drivers/media/v4l2-core/v4l2-ioctl.c:2963
- kernel/bpf/hashtab.c:1382, 1383, 1384, 1574
- net/ipv6/route.c:403
- drivers/bluetooth/btfm_slim.c:307
- drivers/media/rc/bpf-lirc.c:293
- kernel/bpf/btf.c:5456, 5458, 5514
- net/ipv6/mcast.c:1191
- kernel/bpf/net_namespace.c:250
- kernel/params.c:367, 423
- kernel/bpf/cgroup.c:767

+++ net/ipv6/mcast.c
@@ -1188,7 +1188,7 @@ static bool mld_in_v1_mode(const struct inet6_dev *idev)
 		return false;
 	if (mld_in_v1_mode_only(idev))
 		return true;
-	if (idev->mc_v1_seen && time_before(jiffies, idev->mc_v1_seen))
+	if (idev->mc_v1_seen && time_before(jiffies, (unsigned long)idev->mc_v1_seen))
 		return true;
 
 	return false;
